Speedup on force and parallalization and stochastic configurations

+++ Modules/Ensemble.py
@@ -40,7 +40,7 @@ import cellconstructor.Settings
 
 
 import SCHAModules
-import sscha_HP_odd
+#import sscha_HP_odd
 
 # Try to load the parallel library if any
 try:
@@ -148,6 +148,10 @@ class Ensemble:
         
         # A flag that memorize if the ensemble has also the stresses
         self.has_stress = False
+
+
+	# Adding new quantity to speed up the minimization step
+	self.original_v=np.zeros(self.N)
 
     def convert_units(self, new_units):
         """
@@ -322,6 +326,8 @@ class Ensemble:
         self.sscha_forces = np.zeros( (self.N, Nat_sc, 3), order = "F", dtype = np.float64)
         
         self.u_disps = np.zeros( (self.N, Nat_sc * 3), order = "F", dtype = np.float64)
+
+	self.original_v=np.zeros((self.N), dtype=np.float64)
         
         # Add a counter to check if all the stress tensors are present
         count_stress = 0 
@@ -433,7 +439,7 @@ class Ensemble:
             # Allow the garbage collector to free the memory
             del self.stresses
         
-    def save(self, data_dir, population):
+    def save(self, data_dir, population,save_sscha=False):
         """
         SAVE THE ENSEMBLE
         =================
@@ -499,9 +505,13 @@ class Ensemble:
             if save_stress:
                 np.savetxt("%s/pressures_population%d_%d.dat" % (data_dir, population, i+1), self.stresses[i,:,:])
             
-        
+        if save_sscha:
+	    np.savetxt("%s/sscha_energies_supercell_population%d.dat" % (data_dir, population), self.sscha_energies)
+	    for i in xrange(self.N):
+		np.savetxt("%s/sscha_forces_population%d_%d.dat" % (data_dir, population, i+1), self.sscha_forces[i,:,:]) 
 
-    def save_bin(self, data_dir, population_id = 1):
+    def save_bin(self, data_dir, population_id = 1,save_sscha=False):
+
         """
         FAST SAVE OF THE ENSEMBLE
         =========================
@@ -518,7 +528,16 @@ class Ensemble:
                 The id of the population. This can be used to save
                 several ensembles in the same data_dir
         """
+        # Check if the given data_dir is a real directory
+        if os.path.exists(data_dir) and not os.path.isdir(data_dir):
+            raise IOError("Error, the given data_dir %s is not a valid directory." % data_dir)
+        if not os.path.exists(data_dir):
+            os.mkdir(data_dir)
         
+        # Remove the tailoring slash if any
+        if data_dir[-1] == "/":
+            data_dir = data_dir[:-1]
+
         np.save("%s/energies_pop%d.npy" % (data_dir, population_id), self.energies)
         np.save("%s/forces_pop%d.npy" % (data_dir, population_id), self.forces)
         
@@ -529,10 +548,13 @@ class Ensemble:
             np.save("%s/stresses_pop%d.npy" % (data_dir, population_id), self.stresses)
         
         self.dyn_0.save_qe("%s/dyn_gen_pop%d_" % (data_dir, population_id))
+	
+	if save_sscha:
+		np.save("%s/sscha_energies_supercell_population%d.npy" % (data_dir, population_id), self.sscha_energies)
+        	np.save("%s/sscha_forces_pop%d.npy" % (data_dir,population_id), self.sscha_forces)
         
         
-        
-    def load_bin(self, data_dir, population_id = 1, avoid_loading_dyn = False):
+    def load_bin(self, data_dir, population_id = 1, avoid_loading_dyn = False,load_sscha=False):
         """
         LOAD THE BINARY ENSEMBLE
         ========================
@@ -555,7 +577,8 @@ class Ensemble:
         
         # Load the number of configurations
         self.N = len(self.energies)
-        
+        self.original_v=np.zeros((self.N), dtype=np.float64)
+
         stress_path = "%s/stresses_pop%d.npy" % (data_dir, population_id)
         if os.path.exists(stress_path):
             self.stresses = np.load(stress_path)
@@ -573,8 +596,8 @@ class Ensemble:
         super_fc = dyn_supercell.dynmats[0]
         Nat_sc = super_structure.N_atoms
         
-        self.sscha_energies = np.zeros(self.N, dtype = np.float64)
-        self.sscha_forces = np.zeros( (self.N, Nat_sc, 3), order = "F", dtype = np.float64)
+        #self.sscha_energies = np.zeros(self.N, dtype = np.float64)
+        #self.sscha_forces = np.zeros( (self.N, Nat_sc, 3), order = "F", dtype = np.float64)
         self.u_disps = np.zeros( (self.N, 3 * Nat_sc), order = "F", dtype = np.float64)
         
         # Build the structures
@@ -583,12 +606,19 @@ class Ensemble:
             self.structures[i] = super_structure.copy()
             self.structures[i].coords = self.xats[i,:,:]
             self.u_disps[i, :] = (self.xats[i, :, :] - super_structure.coords).reshape( 3*Nat_sc )
-            
-            energy, force = self.dyn_0.get_energy_forces(self.structures[i], supercell = self.supercell, 
+        
+	if load_sscha:
+	    self.sscha_energies=np.load("%s/sscha_energies_supercell_population%d.npy" % (data_dir, population_id))
+	    self.sscha_forces=np.load("%s/sscha_forces_pop%d.npy" % (data_dir,population_id))
+	else:
+	    self.sscha_energies = np.zeros(self.N, dtype = np.float64)
+            self.sscha_forces = np.zeros( (self.N, Nat_sc, 3), order = "F", dtype = np.float64)
+	    for i in xrange(self.N):
+            	energy, force = self.dyn_0.get_energy_forces(self.structures[i], supercell = self.supercell, 
                                                          real_space_fc=super_fc)
             
-            self.sscha_energies[i] = energy
-            self.sscha_forces[i, :, :] = force
+            	self.sscha_energies[i] = energy
+            	self.sscha_forces[i, :, :] = force
 
 
         # Setup the initial weights
@@ -663,6 +693,7 @@ class Ensemble:
         self.rho = np.ones(self.N, dtype = np.float64)
         self.current_dyn = self.dyn_0.Copy()
         self.current_T = self.T0
+	self.original_v=np.zeros(self.N, dtype=np.float64)
         
         
         # Generate the q_start
@@ -761,8 +792,24 @@ class Ensemble:
         self.sscha_energies = sscha_energies_new
         self.sscha_forces = sscha_forces_new
                 
-        
-        
+    def original(self):
+
+	#self.current_T=original_T    
+        super_dyn= self.dyn_0.GenerateSupercellDyn(self.supercell)
+        #w, pols = super_dyn.DyagDinQ(0)
+	Nat_sc = super_dyn.structure.N_atoms
+	disps=np.zeros(np.shape(self.u_disps), dtype=np.double)
+	
+	
+	for i in xrange(self.N):
+	    disps[i,:]=(self.xats[i,:,:]-super_dyn.structure.coords).reshape(3*Nat_sc)
+
+	Ups=np.real(super_dyn.GetUpsilonMatrix(self.T0))
+	
+	for i in xrange(self.N):
+	    self.original_v[i] = disps[i, :].dot(Ups.dot(disps[i, :])) * __A_TO_BOHR__**2	
+
+
     def update_weights(self, new_dynamical_matrix, newT, update_q = False):
         """
         IMPORTANCE SAMPLING
@@ -819,11 +866,23 @@ class Ensemble:
         for i in xrange(self.N):
             self.u_disps[i, :] = (self.xats[i, :, :] - super_structure.coords).reshape( 3*Nat_sc )
             
-            old_disps[i,:] = (self.xats[i, :, :] - super_dyn.structure.coords).reshape( 3*Nat_sc )
-            
+            #old_disps[i,:] = (self.xats[i, :, :] - super_dyn.structure.coords).reshape( 3*Nat_sc )
+
+	    #calculate enegy and forces in fortran
+	    #self.sscha_energies[i]=SCHAModules.harmonic.get_harmonic_energy_from_fc(new_super_dyn.dynmats[0],self.u_disps[i,:])*__A_TO_BOHR__**2
+	    #self.sscha_forces[i,:,:]=(SCHAModules.harmonic.get_harmonic_force_from_fc(new_super_dyn.dynmats[0],self.u_disps[i,:],Nat_sc)).reshape(Nat_sc,3)*__A_TO_BOHR__**2            
+		
+	    
+
             # TODO: this method recomputes the displacements, it is useless since we already have them in self.u_disps
-            self.sscha_energies[i], self.sscha_forces[i, :,:] = new_super_dyn.get_energy_forces(self.structures[i], real_space_fc = new_super_dyn.dynmats[0], displacement = self.u_disps[i, :])
-        t4 = time.time()
+            #self.sscha_energies[i], self.sscha_forces[i, :,:] = new_super_dyn.get_energy_forces(self.structures[i], real_space_fc = new_super_dyn.dynmats[0], displacement = self.u_disps[i, :])
+        
+
+	#super_fast
+	forces,self.sscha_energies =SCHAModules.harmonic.update(new_super_dyn.dynmats[0],self.u_disps,Nat_sc,self.N)
+	self.sscha_forces=forces.reshape(self.N,Nat_sc,3)
+	t4 = time.time()
+
 
         if update_q:
             self.current_q = CC.Manipulate.GetQ_vectors(self.structures, new_super_dyn, self.u_disps) 
@@ -846,8 +905,10 @@ class Ensemble:
         
         # Get the covariance matrices of the ensemble
         ups_new = np.real(new_super_dyn.GetUpsilonMatrix(self.current_T))
-        ups_old = np.real(super_dyn.GetUpsilonMatrix(self.T0))
+        #ups_old = np.real(super_dyn.GetUpsilonMatrix(self.T0))
+	
 
+	
         # Get the normalization ratio
         #norm = np.sqrt(np.abs(np.linalg.det(ups_new) / np.linalg.det(ups_old))) 
         norm = np.prod( old_a / new_a)
@@ -860,10 +921,15 @@ class Ensemble:
         rho_tmp = np.ones( self.N, dtype = np.float64) * norm 
         if __DEBUG_RHO__:
             print("Norm factor:", norm)
-        for i in xrange(self.N):
-            v_new = self.u_disps[i, :].dot(ups_new.dot(self.u_disps[i, :])) * __A_TO_BOHR__**2
-            v_old = old_disps[i, :].dot(ups_old.dot(old_disps[i, :])) * __A_TO_BOHR__**2
-            rho_tmp[i] *= np.exp(-0.5 * (v_new - v_old) )
+        #for i in xrange(self.N):
+            #v_new = self.u_disps[i, :].dot(ups_new.dot(self.u_disps[i, :])) * __A_TO_BOHR__**2
+            #v_old = old_disps[i, :].dot(ups_old.dot(old_disps[i, :])) * __A_TO_BOHR__**2
+            #rho_tmp[i] *= np.exp(-0.5 * (v_new - v_old) )
+
+	for i in xrange(self.N):    
+	    v_new=self.u_disps[i,:].dot(ups_new.dot(self.u_disps[i,:]))* __A_TO_BOHR__**2
+	    rho_tmp[i] *= np.exp(-0.5* (v_new-self.original_v[i]))
+
         # Lets try to use this one
         self.rho = rho_tmp
 
@@ -1082,6 +1148,38 @@ class Ensemble:
             return free_energy, error
         return free_energy
 
+    def get_entropy(self):
+	"""
+	Obtain the entropy for the system.
+	Done by taking minus the derivative of the Free energy with respect to the temperature.
+
+	The result is in Ry/K
+
+	.. math ::
+
+		\\S = - \\frac{d F}{d T} = -k_B \\sum_\mu ln(1- e^{-\beta \hbar \omega_\mu}) + \sum_\mu \\frac{\hbar}{T} n_\mu \omega_\mu
+
+        Where n_\mu is the bosonic occupation factor
+
+	"""
+	super_dyn=self.current_dyn.GenerateSupercellDyn(self.supercell)
+
+	w, pols= super_dyn.DyagDinQ(0)
+
+	#Exclude translations
+	w = w[~CC.Methods.get_translations(pols, super_dyn.structure.get_masses_array())]
+
+	#Calculate the bosonic occupation factor
+
+	
+	
+	if self.current_T== 0:
+	    entropy = 0
+	else:
+	    n_bos = 1 / (np.exp(w * __RyToK__ / self.current_T) - 1) 
+	    entropy =  np.sum(- np.log(1- np.exp(-w * __RyToK__ / self.current_T))/__RyToK__ + w *n_bos /self.current_T)
+	
+	return entropy
 
     def get_free_energy_interpolating(self, target_supercell, support_dyn_coarse = None, support_dyn_fine = None, error_on_imaginary_frequency = True, return_error = False):
         """
@@ -1418,7 +1516,7 @@ class Ensemble:
         return cov_mat
     
     
-    def get_stress_tensor(self, offset_stress = None, add_centroid_contrib = False):
+    def get_stress_tensor(self, offset_stress = None, add_centroid_contrib = False,even_odd=False,use_spglib=False):
         """
         GET STRESS TENSOR
         =================
@@ -1437,6 +1535,8 @@ class Ensemble:
             add_centroid_contrib : bool, optional
                 If true the contribution of the centroid is added. This is always zero when
                 the system is relaxed.
+	    use_spglib: bool
+		If true use the spglib library to perform the symmetrization
         
         Results
         -------
@@ -1480,7 +1580,28 @@ class Ensemble:
         
         abinit_stress = np.einsum("abc -> cba", self.stresses, order = "F")
         
-        stress, err_stress = SCHAModules.get_stress_tensor(volume, self.forces / __A_TO_BOHR__, u_disps * __A_TO_BOHR__, 
+	## Introducing the possibility to reduce the error by coupling even-odd configuration
+
+	if even_odd:
+	    # Define the new symmetrized quantities
+	    reduced_N =self.N/2
+	    eo_abinit_stress=np.zeros((3,3,reduced_N))
+	    eo_forces=np.zeros((reduced_N,nat,3))
+	    eo_disps=np.zeros((reduced_N,nat,3))
+	    eo_rho=np.zeros(reduced_N)
+		
+	    # Calculate them		
+	    for i in range(reduced_N):
+                eo_abinit_stress[:,:,i]=(abinit_stress[:,:,2*i]+abinit_stress[:,:,2*i+1])/2
+		eo_forces[i,:,:]=(self.forces[2*i,:,:]-self.forces[2*i+1,:,:])/2
+		eo_disps[i,:,:]=u_disps[2*i,:,:]
+		eo_rho[i]=self.rho[2*i]
+	
+	    stress, err_stress = SCHAModules.get_stress_tensor(volume, eo_forces / __A_TO_BOHR__, eo_disps * __A_TO_BOHR__, 
+                                                           eo_abinit_stress, wr, er, self.current_T, eo_rho, "err_yesrho", 
+                                                           reduced_N, nat, len(wr))
+        else:
+            stress, err_stress = SCHAModules.get_stress_tensor(volume, self.forces / __A_TO_BOHR__, u_disps * __A_TO_BOHR__, 
                                                            abinit_stress, wr, er, self.current_T, self.rho, "err_yesrho", 
                                                            self.N, nat, len(wr))    
         
@@ -1539,6 +1660,10 @@ class Ensemble:
 
         # Symmetrize the stress tensor
         qe_sym = CC.symmetries.QE_Symmetry(self.current_dyn.structure)
+	if not use_spglib:
+	    qe_sym.SetupQPoint()
+	else:
+	    qe_sym.SetupFromSPGLIB()
         qe_sym.ApplySymmetryToMatrix(stress, err_stress)
         
         return stress, err_stress
@@ -2388,7 +2513,7 @@ class Ensemble:
                 else:
                     # Lets call the C code with openMP support
                     # to compute the d3 faster
-                    sscha_HP_odd.GetV3(X, Y, n_modes_sc, self.N, d3)
+                    #sscha_HP_odd.GetV3(X, Y, n_modes_sc, self.N, d3)
                     d3 *= self.N / N_eff
             else:
                 d3 = load_d3
@@ -2592,7 +2717,9 @@ class Ensemble:
                 odd_corr = odd_corrs[i_freq]
             cart_odd_corr = np.einsum("ai,bj,ij -> ab", pols_sc, pols_sc, odd_corr)
             cart_odd_corr *= np.outer(np.sqrt(_m_),  np.sqrt(_m_))
-
+	
+	    if save_all:
+                np.save("odd_corr_%d.npy" % i_freq, cart_odd_corr)
 
             total_self_energy = cart_odd_corr 
             if not return_only_correction:
@@ -2603,8 +2730,7 @@ class Ensemble:
             if N_w >1:
                 total_self_energies.append(total_self_energy)
             
-            if save_all:
-                np.save("odd_corr_%d.npy" % i_freq, cart_odd_corr)
+            
         
         # Return the new dynamical matrix in the supercell
         if N_w > 1:
@@ -2684,11 +2810,9 @@ class Ensemble:
             structures = comm.bcast(self.structures, root = 0)            
             nat3 = comm.bcast(self.current_dyn.structure.N_atoms* 3* np.prod(self.supercell), root = 0)
             N_rand = comm.bcast(self.N, root=0)
-            
             # Setup the label of the calculator
             ase_calculator = comm.bcast(ase_calculator, root = 0)
             ase_calculator.set_label("esp_%d" % rank) # Avoid overwriting the same file
-            
             compute_stress = comm.bcast(compute_stress, root = 0)
             
             # Check if the parallelization is correct        
@@ -2703,7 +2827,7 @@ class Ensemble:
             N_rand = self.N
             
         # Only for the master
-        
+  
         # Prepare the energy, forces and stress array
         energies = np.zeros( N_rand / size, dtype = np.float64)
         forces = np.zeros( ( N_rand / size) * nat3 , dtype = np.float64)
@@ -2740,6 +2864,11 @@ class Ensemble:
                     print ("Rerun the job %d" % i)
                     count_fails += 1
                     if count_fails >= 5:
+                        #Anyway save the partial results
+         		np.save("%s/partial_energies_pop%d.npy" % (data_dir, population_id), self.energies)
+   		        np.save("%s/partial_forces_pop%d.npy" % (data_dir, population_id), self.forces)
+        	        if self.has_stress:
+            		    np.save("%s/partla_stresses_pop%d.npy" % (data_dir, population_id), self.stresses)
                         run = False
                         raise ValueError("Error in the ASE calculator for more than 5 times")
             
@@ -2786,6 +2915,52 @@ class Ensemble:
         else:
             self.has_stress = False
             
-            
-        
+    
+    def get_kinetic_energy(self, element=None):
+	"""
+	This function returns the kinetic energy per atom of the system in meV.
+	Keep in mind that this formula is correct only at the minimum of the free energy (and at the equilibrium volume??)
+
+	K= \frac{1}{2} \sum_{\mu s} \hbar \omega_mu (n_\mu +\frac{1}{2}) \sum_j \epsilon_\mu^{sj} \epsilon_\mu^{sj}
+	Parametes
+	----------	
+		element :
+			Allows to isolate the kinetic energy of a single element        
+	
+
+	"""
+ 	super_dyn=self.current_dyn.GenerateSupercellDyn(self.supercell)
+	nat_sc=super_dyn.structure.N_atoms
+
+	w, pols= super_dyn.DyagDinQ(0)
+	#Exclude translations
+	w = w[~CC.Methods.get_translations(pols, super_dyn.structure.get_masses_array())]
+	pols = pols[ :, ~CC.Methods.get_translations(pols, super_dyn.structure.get_masses_array()) ]
+	n_modes=len(w)
+	#Calculate the bosonic occupation factor
+
+	if self.current_T== 0:
+	    n_bos = 0
+	else:
+	    n_bos = 1 / (np.exp(w * __RyToK__ / self.current_T) - 1) 
+
+	#Calculate the potential energy and reshape the polarization factor
+	pot_energy=(n_bos+0.5)*w*13605.698066
+	pol_vec=pols.reshape(nat_sc,3, n_modes)
+	kinetic_energy=0
+
+	if element is None:
+	    return np.sum(pot_energy)/(2*nat_sc)
+	else:
+	    masses=super_dyn.structure.get_masses_array()
+	    atom_symbol=super_dyn.structure.atoms
+	    for i in range(n_modes):
+		num_of_atoms= 0
+	        for j in range(nat_sc):
+	            if atom_symbol[j]==element:
+			num_of_atoms += 1
+			kinetic_energy += pot_energy[i]*np.sum(pol_vec[j,:,i]*np.conj(pol_vec[j,:,i]))/2
+	    return kinetic_energy/num_of_atoms		
+
+
